:recycle: Type 파일 분리
