+++ libs/shared/common-layout/type-common-layout.ts
@@ -1,0 +1,9 @@
+interface UiCommonLayoutProps {
+  title: string
+  titleAlign?: 'start' | 'center'
+  titleSize?: number
+  gap?: number
+  children: React.ReactNode
+}
+
+export type { UiCommonLayoutProps }
