+++ libs/shared/common-modal/type-common-modal.ts
@@ -1,0 +1,9 @@
+interface UiCommonModalProps {
+  padding?: string | number
+  borderRadius?: number
+  backgroundColor?: 'White' | 'Gray'
+  closeButtonSize?: number
+  children: React.ReactNode
+}
+
+export type { UiCommonModalProps }
